PixelPlaceBot v1.0
This simple bot draws the coordinates that you provided automatically. There isn't any GUI right now, but GUI will come at newer versions.
Guideline:
Add the coordinates that you want to draw as draw(456,215) at the end of the code.
After scrolling 8 times, you need to drag to the desired area before hitting the draw button. So, the area that you want to draw should be visible before hitting at the draw button. It will be fixed at the newer versions.
Run the program and follow the guides.
Fail-safe will be triggered if you drag your mouse at the top right of the screen.